Clean up geometry

+++ visualization-client/src/view/mod.rs
@@ -43,12 +43,10 @@ impl CanvasView {
         self.context.begin_path();
 
         let first_vertex = &object.shape.vertices[0];
-        self.context
-            .move_to(f64::from(first_vertex.x), f64::from(first_vertex.y));
+        self.context.move_to(first_vertex.x, first_vertex.y);
 
         for vertex in &object.shape.vertices[1..] {
-            self.context
-                .line_to(f64::from(vertex.x), f64::from(vertex.y));
+            self.context.line_to(vertex.x, vertex.y);
         }
 
         self.context.close_path();
